forpython/send_notification/requirements.txt
click==7.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Flask==1.1.2
python-telegram-bot
